refactor: move LogEventHVL to a header file
