+++ chrono_common/StoryChunkWriter.h
@@ -6,6 +6,7 @@
 #include <H5Cpp.h>
 #include "chrono_monitor.h"
 #include "StoryChunk.h"
+#include "LogEventHVL.h"
 
 namespace chronolog
 {
